Set float32 defaults; added log return to predict data.

+++ alphai_feature_generation/feature/feature.py
@@ -376,7 +376,7 @@ class FinancialFeature(object):
 
         variances[variances == 0] = 1.0  # FIXME Hack
 
-        diag_cov_matrix = np.diag(variances)
+        diag_cov_matrix = np.diag(variances, dtype=np.float32)
         return means, diag_cov_matrix
